requirements.txt
typer
pydantic
requests
pandas
playwright 
python-dotenv
selenium
webdriver-manager
fastapi
gunicorn
uvicorn
sqlalchemy